Updating test helper to more closely match ActiveMerchant's
This updates the test helper to more closely match that found in the ActiveMerchant gem. By doing this, it reverts some changes to the test classes, choosing to use test unit instead of minitest. A number of dependencies have also been removed as they don't appear to be used by the tests.

Also, this fixes an incorrect loading of all_fixtures that was previously introduced to resolve a linting issue.